// src/constants/enums.tsx
import React from 'react'

export enum SERVICE {
    AGUA_POTABLE_Y_DRENAJE = 'Agua Potable y Drenaje',
    SEGURIDAD_PUBLICA = 'Seguridad Publica',
    PROTECCION_CIVIL = 'Protección civil',
    MANTENIMIENTO_CALLES = 'Mantenimiento de Calles/ Baches',
    ALUMBRADO_PUBLICO = 'Alumbrado PUblico',
    BASURA = 'Basura',
    VIOLENCIA_DE_GENERO = 'Violencia de genero o contra la mujer',
    MALTRATO_ANIMAL = 'Maltrato Animal',
    MANTENIMIENTO_URBANO = 'Mantenimiento Urbano y limpia',
    ALERTA_SISMICA = 'Alerta sismica',
}